requirements.txt
fastapi==0.115.8
mistralai==1.5.0
openai==1.62.0
orjson==3.10.15
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0